Remove cosineSilhouette implementation and move squaredSilhouette to ClusteringEvaluator
